### 2025-05-29 18d8263
Add JOSS submitted icon/link to readme
