### 2025-05-29 4b119ab
force gnu17 during compilation
