+++ imgmarker/io.py
@@ -1,4 +1,4 @@
-"""Image Marker's I/O module containing functions for loading and saving data."""
+"""Image Marker's I/O module containing functions for reading and saving data."""
 
 import os
 import numpy as np
@@ -10,6 +10,336 @@ from typing import Tuple, List
 import csv
 import datetime as dt
 
+
+class MarkFile:
+    def __init__(self,path):
+        self.path = path
+    
+    def __eq__(self, value):
+        if hasattr(value,'path'):
+            return self.path == value.path
+        else:
+            return self.path == value
+        
+    def read(self,images:List[image.Image]) -> Tuple[List[image.Image],List[Mark]]:
+        """
+        Takes data from marks.csv and images.csv and from them returns a list of `imgmarker.image.Image`
+        objects.
+
+        Returns
+        ----------
+        images: list[`imgmarker.image.Image`]
+        """
+
+        imageless = []
+        
+        # Get list of marks for each image
+        if os.path.exists(self.path):
+            with open(self.path,'r') as f:
+                delimiter = '|' if '|' in f.readline() else ','
+                f.seek(0)
+                reader = csv.DictReader(f,delimiter=delimiter)
+
+                for row in reader:
+                    keys = row.copy().keys()
+                    for key in keys: row[key.strip().lower()] = row.pop(key).strip()
+
+                    # Default values
+                    name = 'None'
+                    group = config.GROUP_NAMES.index('None')
+                    shape = 'rect'
+                    label = 'None'
+                    x,y = nan,nan
+                    ra,dec = nan,nan
+                    size = None
+                    size_unit = None
+
+                    # Values from row
+                    if 'date' in row: date = row['date']
+
+                    if 'image' in row: name = row['image']
+
+                    if 'group' in row: 
+                        group = config.GROUP_NAMES.index(row['group'])
+                        shape = config.GROUP_SHAPES[group]
+                        
+                    if 'label' in row: label = row['label']
+
+                    if 'x' in row: x = float(row['x'])
+                    if 'y' in row: y = float(row['y'])
+
+                    if 'ra' in row: ra = float(row['ra'])
+                    if 'dec' in row: dec = float(row['dec'])
+
+                    if 'size(arcsec)' in row: 
+                        size = float(row['size(arcsec)'])
+                        size_unit = 'arcsec'
+                    
+                    if 'size(px)' in row:
+                        size = float(row['size(px)'])
+                        size_unit = 'px'
+
+                    if 'size' in row:
+                        size = float(row['size'])
+                        size_unit = 'px'
+
+                    if name != 'None':
+                        for img in images:
+                            if (name == img.name) and (not isnan(float(x))) and (not isnan(float(y))):
+                                args = (float(x),float(y))
+                                kwargs = {'image': img, 'group': group, 'shape': shape}
+
+                                if label != 'None': 
+                                    kwargs['text'] = label
+
+                                if size != None: 
+                                    kwargs['size'] = size
+                                    kwargs['size_unit'] = size_unit
+
+                                mark = Mark(*args, **kwargs)
+                                mark.dst = self.path
+                                img.marks.append(mark)
+
+                    else:
+                        if isnan(float(ra)) and isnan(float(dec)):
+                            args = (float(x),float(y))
+                            kwargs = {'image': None, 'group': group, 'shape': shape}
+                        else:
+                            args = ()
+                            kwargs = {'image': None,'group': group, 'shape': shape, 'ra': ra, 'dec': dec}
+
+                        if label != 'None': 
+                            kwargs['text'] = label
+
+                        if size != None: 
+                            kwargs['size'] = size
[…]
