### 2025-05-29 585dccf
fixed coordinates not included during pip install; fixed Angle bug in update_pos
