### 2025-05-31 8ac014f
fixed speed issues with loading mark files; fixed some sizing issues with marks and labels
