+++ docs/source/reference/api/imgmarker.rst
@@ -7,6 +7,7 @@ Subpackages
 .. toctree::
    :maxdepth: 4
 
+   imgmarker.coordinates
    imgmarker.gui
    imgmarker.image
 
