### 2025-05-31 8944fd7
fixed saving pixel coordinates
