### 2025-05-31 bd26352
fixed some speed issues during saving; added editing colors of each mark file individually
