+++ imgmarker/gui/window.py
@@ -8,7 +8,8 @@ from imgmarker.gui.pyqt import (
     QLineEdit, QFileDialog, QIcon, QFont, QAction, 
     Qt, QPoint, QSpinBox, QMessageBox, QTableWidget, 
     QTableWidgetItem, QHeaderView, QShortcut,
-    QDesktopServices, QUrl, QMenu, QColorDialog, PYQT_VERSION_STR
+    QDesktopServices, QUrl, QMenu, QColorDialog, 
+    QPen, QBrush, QPixmap, QPainter, PYQT_VERSION_STR
 )
 from imgmarker.gui import Screen, QHLine, PosWidget, RestrictedLineEdit, DefaultDialog
 from imgmarker import HEART_SOLID, HEART_CLEAR, __version__, __license__, __docsurl__
@@ -461,10 +462,19 @@ class MarkMenu(QMenu):
 
         self.menus[path].addSeparator()
 
+        color_action = QAction('Default Color...', self)
+        color_action.triggered.connect(partial(self.mainwindow.update_colors,path))
+        color_action.setToolTip('Edit color of marks that aren\'t part of a group')
+        self.menus[path].addAction(color_action)
+        self.update_color(path)
+
+        self.menus[path].addSeparator()
+
         if path == self.mainwindow.markfile.path:
             labels_action.setShortcuts(['Ctrl+l'])
 
             del_marks_action = QAction(f'Delete Marks in Current Image', self)
+            
             del_marks_action.triggered.connect(partial(self.mainwindow.del_usermarks,'all'))
             self.menus[path].addAction(del_marks_action)
         else:
@@ -483,6 +493,22 @@ class MarkMenu(QMenu):
         else:
             self.marks_action(path).setEnabled(True)
             self.labels_action(path).setEnabled(True)
+
+    def update_color(self,path):
+        s = 14
+        pixmap = QPixmap(s,s)
+        painter = QPainter(pixmap)
+        
+        pen = QPen(Qt.GlobalColor.black, 2)
+        painter.setBrush(config.DEFAULT_COLORS[path])
+        painter.setPen(pen)
+        painter.drawRect(0, 0, s, s)
+        painter.end()
+        icon = QIcon(pixmap)
+        self.color_action(path).setIcon(icon)
+
+    def color_action(self,path):
+        return [action for action in self.menus[path].actions() if action.text() == "Default Color..."][0]
     
     def marks_action(self,path):
         return [action for action in self.menus[path].actions() if action.text() == "&Show Marks"][0]
@@ -676,13 +702,6 @@ class MainWindow(QMainWindow):
         redo_mark_action.setShortcuts(['Ctrl+Shift+z'])
         redo_mark_action.triggered.connect(self.redo_prev_mark)
         edit_menu.addAction(redo_mark_action)
-
-        edit_menu.addSeparator()
-
-        color_action = QAction('&Default Color...', self)
-        color_action.triggered.connect(self.update_colors)
-        color_action.setToolTip('Edit color of marks that aren\'t part of a group')
-        edit_menu.addAction(color_action)
 
         ### Settings menu
         edit_menu.addSeparator()
@@ -940,6 +959,7 @@ class MainWindow(QMainWindow):
 
     def closeEvent(self, a0):
         self.update_comments()
+        self.save()
         self.about_window.close()
         self.blur_window.close()
         self.frame_window.close()
@@ -1060,27 +1080,9 @@ class MainWindow(QMainWindow):
         mark_dst = shutil.copy(src,dst)
 
         self.images, imageless_marks = io.MarkFile(mark_dst).read(self.images)
-            
         self.imageless_marks += imageless_marks
 
         self.update_marks()
-
-        '''if catalog and not test:
-            self.color_picker_window = ColorPickerWindow(self)
[…]
