### 2025-05-31 7f822c8
changed order of blurring and rescaling; fixed parameter n in gaussian_filter not being used
