### 2025-06-09 7160319
Update marks.rst
Fixed small typos, repeated bullet, and added delete/select marks info
