### 2025-06-09 1004b53
Update controls.rst
