### 2025-06-09 479cb1e
Fixed controls window for deleting marks and typo
