+++ imgmarker/gui/window.py
@@ -342,11 +342,11 @@ class ControlsWindow(QWidget):
         
     def update_text(self):
         # Lists for keybindings
-        actions_list = ['Next','Back','Change frame','Delete','Enter comment', 'Focus', 'Zoom in/out', 'Copy mark coordinates' ,'Favorite']
+        actions_list = ['Next','Back','Change frame','Delete mark','Delete selected marks','Enter comment', 'Focus', 'Zoom in/out', 'Copy mark coordinates' ,'Favorite']
         group_list = [f'Group \"{group}\"' for group in config.GROUP_NAMES[1:]]
         category_list = [f'Category \"{category}\"' for category in config.CATEGORY_NAMES[1:]]
         actions_list = group_list + category_list + actions_list
-        buttons_list = ['1 OR Left click', '2', '3', '4', '5', '6', '7', '8', '9', 'Ctrl+1', 'Ctrl+2', 'Ctrl+3', 'Ctrl+4', 'Ctrl+5', 'Tab', 'Shift+Tab', 'Spacebar', 'Delete OR Shift+Right Click', 'Enter', 'Middle Click', 'Scroll Wheel', 'Ctrl + C', 'F']
+        buttons_list = ['1 OR Left Click', '2', '3', '4', '5', '6', '7', '8', '9', 'Ctrl+1', 'Ctrl+2', 'Ctrl+3', 'Ctrl+4', 'Ctrl+5', 'Tab', 'Shift+Tab', 'Spacebar', 'Shift+Left Click', 'Delete', 'Enter', 'Middle Click', 'Scroll Wheel', 'Ctrl + C', 'F']
         
         items = [ (action, button) for action, button in zip(actions_list, buttons_list) ]
 
