### 2025-06-09 1fbce35
fixed crashes when copying WCS coordinates
