### 2025-06-09 88e9e72
Added all checks for disabling show/hide marks if undone/redone, and fixed undone/redone with mark limits
