### 2025-06-09 31fff6d
removed erroneous use of 'value' attribute for Angle objects
