### 2025-06-10 e03e8be
Catch exception when importing invalid mark file
