### 2025-06-10 277cfb3
Translated favorites file saving/reading into ImagesFile format
