+++ imgmarker/io.py
@@ -10,7 +10,6 @@ from typing import Tuple, List
 import csv
 import datetime as dt
 
-
 class MarkFile:
     def __init__(self,path:str):
         self.path = path
@@ -260,7 +259,6 @@ class MarkFile:
             for row in rows:
                 writer.writerow(row)
 
-
 class ImagesFile:
     def __init__(self):
         self.path = os.path.join(config.SAVE_DIR,f'{config.USER}_images.csv')
@@ -314,9 +312,6 @@ class ImagesFile:
 
         Parameters
         ----------
-        date: str
-            A string containing the current date in ISO 8601 extended format.
-
         images: list[`imgmarker.image.Image`]
             A list of Image objects for each image from the specified image directory.
 
@@ -351,8 +346,93 @@ class ImagesFile:
             writer.writeheader()
             for row in image_rows:
                 writer.writerow(row)
-        
 
+class FavoritesFile:
+    def __init__(self):
+        self.path = os.path.join(config.SAVE_DIR,f'{config.USER}_favorites.csv')
+    
+    def __eq__(self, value):
+        if hasattr(value,'path'):
+            return self.path == value.path
+        else:
+            return self.path == value
+        
+    def read(self) -> List[str]:
+        """
+        Takes data from favorites.csv and from them returns a list of image file names
+        with full directory.
+
+        Returns
+        ----------
+        favorites: list[`str`]
+        """
+
+        favorites:List[str] = []
+        
+        # Get list of images from favorites.csv
+        if os.path.exists(self.path):
+            with open(self.path,'r') as f:
+                delimiter = '|' if '|' in f.readline() else ','
+                f.seek(0)
+                reader = csv.DictReader(f,delimiter=delimiter)
+                
+                for row in reader:
+                    keys = row.copy().keys()
+                    for key in keys: row[key.strip().lower()] = row.pop(key).strip()
+                    date,name,categories,comment = row['date'], row['image'], row['categories'], row['comment']
+
+                    favorites.append(name)
+        return favorites
+
+    def save(self, favorites:List[str], images:List['image.Image']):
+        """
+        Saves favorites data.
+
+        Parameters
+        ----------
+        favorites: list[str]
+            A list of strings of each image file name from the specified image directory.
+
+        images: list['image.Image']
+            A list of image objects.
+            
+        Returns
+        ----------
+        None
+        """
+        # fav_out_path = os.path.join(config.SAVE_DIR, f'{config.USER}_favorites.csv')
+        date = dt.datetime.now(dt.timezone.utc).date().isoformat()
+        image_rows:list[dict] = []
+    
+        favorited = [favorite for favorite in images if favorite.name in favorites]
+
+        for img in favorited:
+            if img.seen:
+                name = img.name
+                comment = img.comment
+
+                category_list = img.categories
+                category_list.sort()
+                if (len(category_list) != 0):
+                    categories = '+'.join([config.CATEGORY_NAMES[i] for i in category_list])
+                else: categories = 'None'
+
+                image_rows.append({'date': str(date),
+                                    'image': str(name),
+                                    'RA': str(img.wcs_center[0]),
+                                    'DEC': str(img.wcs_center[1]),
+                                    'categories': str(categories),
+                                    'comment': str(comment)})
+        if len(favorited) > 0:
+            with open(self.path, 'w') as f:
[…]
