### 2025-06-10 90c9389
Update controls.rst
