### 2025-06-10 cb89c1f
Updated controls window
