### 2025-06-10 9ad1774
skip over files in imports folder that aren't valid mark files
