### 2025-06-12 9dccf1a
fixed markfile, imagesfile, and favoritesfile attributes not being updated when opening a new save
