### 2025-06-12 833e660
fixed some shortcut issues
