+++ imgmarker/gui/window.py
@@ -429,26 +429,25 @@ class MarkMenu(QMenu):
         super().__init__()
         self.menus:dict[str,QMenu] = {}
         self.mainwindow = mainwindow
-        self.setTitle('&Mark')
+        self.setTitle('Mark')
 
     def menu_setup(self,path:str):
         file = path.split(os.sep)[-1]
 
         if path == self.mainwindow.markfile.path:
-            self.menus[path] = QMenu(f'&{file} (default)')
+            self.menus[path] = QMenu(f'{file} (default)')
         else:
-            self.menus[path] = QMenu(f'&{file}')
+            self.menus[path] = QMenu(f'{file}')
 
         # Toggle marks
-        marks_action = QAction('&Show Marks', self)
-        marks_action.setShortcuts(['Ctrl+m'])
+        marks_action = QAction('Show Marks', self)
         marks_action.setCheckable(True)
         marks_action.setChecked(True)
         marks_action.triggered.connect(partial(self.mainwindow.toggle_marks,path))
         self.menus[path].addAction(marks_action)
         
         ### Toggle mark labels menu
-        labels_action = QAction('&Show Mark Labels', self)
+        labels_action = QAction('Show Mark Labels', self)
         labels_action.setCheckable(True)
         labels_action.setChecked(True)
         labels_action.triggered.connect(partial(self.mainwindow.toggle_mark_labels,path))
@@ -473,14 +472,14 @@ class MarkMenu(QMenu):
 
         if path == self.mainwindow.markfile.path:
             labels_action.setShortcuts(['Ctrl+l'])
+            marks_action.setShortcuts(['Ctrl+m'])
 
             del_marks_action = QAction(f'Delete Marks in Current Image', self)
             
             del_marks_action.triggered.connect(partial(self.mainwindow.del_usermarks,'all'))
             self.menus[path].addAction(del_marks_action)
-        else:
-            labels_action.setShortcuts(['Ctrl+Shift+l'])
 
+        else:
             del_file_action = QAction(f'Delete', self)
             del_file_action.triggered.connect(partial(self.mainwindow.del_markfile,path))
             self.menus[path].addAction(del_file_action)
@@ -512,10 +511,10 @@ class MarkMenu(QMenu):
         return [action for action in self.menus[path].actions() if action.text() == "Default Color..."][0]
     
     def marks_action(self,path):
-        return [action for action in self.menus[path].actions() if action.text() == "&Show Marks"][0]
+        return [action for action in self.menus[path].actions() if action.text() == "Show Marks"][0]
     
     def labels_action(self,path):
-        return [action for action in self.menus[path].actions() if action.text() == "&Show Mark Labels"][0]
+        return [action for action in self.menus[path].actions() if action.text() == "Show Mark Labels"][0]
         
 class MainWindow(QMainWindow):
     """Class for the main window."""
@@ -660,16 +659,16 @@ class MainWindow(QMainWindow):
         menubar = self.menuBar()
 
         ## File menu
-        file_menu = menubar.addMenu("&File")
+        file_menu = menubar.addMenu("File")
 
         ### Open file menu
-        open_action = QAction('&Open Save...', self)
+        open_action = QAction('Open Save...', self)
         open_action.setShortcuts(['Ctrl+o'])
         open_action.triggered.connect(self.open)
         file_menu.addAction(open_action)
 
         ### Open new image folder menu
-        import_ims_action = QAction('&Open Images...', self)
+        import_ims_action = QAction('Open Images...', self)
         import_ims_action.setShortcuts(['Ctrl+Shift+i'])
[…]
