### 2025-06-12 77bf802
disable editing default color if mark file does not contain any imageless marks
