### 2025-06-12 e04a3f2
Added warning if limit is set above number of placed marks
