### 2025-06-17 ccc65df
Fixed over limit popup showing when closing imgmarker if limit too low
