### 2025-06-17 8bf2443
added indicator that image has already been seen
