### 2025-06-25 29ba12b
Updated LICENSE and added copyright line to each source code file
