+++ imgmarker/image/convolution.py
@@ -1,3 +1,9 @@
+"""
+Copyright © 2025, UChicago Argonne, LLC
+
+Full license found at _YOUR_INSTALLATION_DIRECTORY_/imgmarker/LICENSE
+"""
+
 import numpy as np
 from scipy.ndimage import uniform_filter
 
