+++ imgmarker/__init__.py
@@ -4,7 +4,7 @@ Copyright © 2025, UChicago Argonne, LLC
 Full license found at _YOUR_INSTALLATION_DIRECTORY_/imgmarker/LICENSE
 """
 
-__version__ = '1.1.1'
+__version__ = '1.1.2'
 __license__ = 'MIT License'
 __docsurl__ = 'https://imgmarker.readthedocs.io/en/latest/'
 
