### 2025-07-01 817e7ca
Implement additional keyboard shortcuts
- Add "z" shortcut to check category 1 (flag1_true)
- Add "x" shortcut to uncheck category 1 (flag1_false)
- Add Left/Right arrow shortcuts for image navigation
