### 2025-07-01 bacfc8d
Fix syntax error where literal quote was being used inside an f-string.
