### 2025-07-02 808d577
fixed pyinstaller .spec file for mac
