+++ .github/workflows/build.yml
@@ -31,7 +31,7 @@ jobs:
       fail-fast: false
       matrix:
         include:
-          - os: windows-2019
+          - os: windows-2022
             spec: win
             arch: x64
 
