### 2025-07-07 2b25621
Added arXiv badge and link
