+++ README.rst
@@ -21,10 +21,13 @@
 .. |JOSS| image:: https://joss.theoj.org/papers/21a640a3fd4e2ad1815891a61af4a21b/status.svg
     :target: https://joss.theoj.org/papers/21a640a3fd4e2ad1815891a61af4a21b
 
+.. |arxiv| image:: https://img.shields.io/badge/arXiv-2507.02153-b31b1b.svg
+    :target: https://arxiv.org/abs/2507.02153
+
 |icon| Image Marker
 #####################
 
-|github| |license| |pyqt| |docs| |CI| |JOSS|
+|github| |license| |pyqt| |docs| |CI| |JOSS| |arxiv|
 
 Image Marker is a tool for marking, categorizing, and annotating TIFF, FITS, PNG, and JPEG files.
 Image Marker does not modify the image itself, it only displays it and creates text files containing 
