### 2025-07-29 55ada77
show 'Cmd' instead of 'Ctrl' in controls menu when on mac
