+++ imgmarker/__init__.py
@@ -5,6 +5,7 @@ __docsurl__ = 'https://imgmarker.readthedocs.io/en/latest/'
 import sys
 import os
 from importlib import resources
+from platform import system
 
 def resource_path(resource):
     if hasattr(sys,'_MEIPASS'):
@@ -12,6 +13,7 @@ def resource_path(resource):
     else: 
         return str(resources.files(__package__).joinpath(resource))
 
+OS = system()
 ICON = resource_path('icon.ico')
 HEART_SOLID = resource_path('heart_solid.ico')
 HEART_CLEAR = resource_path('heart_clear.ico')
