### 2025-09-29 2fd80f0
Update paper.md
