### 2025-09-29 c221586
update abbreviations in bib
