+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{Faber_M-L,
        author = {{Faber}, S.~M. and {Jackson}, R.~E.},
         title = "{Velocity dispersions and mass-to-light ratios for elliptical galaxies.}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {Elliptical Galaxies, Galactic Nuclei, Mass To Light Ratios, Power Spectra, Velocity Distribution, Astronomical Spectroscopy, Data Reduction, Light (Visible Radiation), M Stars, Noise Spectra, Spectral Line Width, Astrophysics},
          year = 1976,
         month = mar,
@@ -15,7 +15,7 @@
 @ARTICLE{Rozo14,
        author = {{Rozo}, E. and {Rykoff}, E.~S.},
         title = "{redMaPPer II: X-Ray and SZ Performance Benchmarks for the SDSS Catalog}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {galaxies: clusters: general, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2014,
         month = mar,
@@ -35,7 +35,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Sobrin22,
        author = {{Sobrin}, J.~A. and {Anderson}, A.~J. and {Bender}, A.~N. and {Benson}, B.~A. and {Dutcher}, D. and {Foster}, A. and {Goeckner-Wald}, N. and {Montgomery}, J. and {Nadolski}, A. and {Rahlin}, A. and {Ade}, P.~A.~R. and {Ahmed}, Z. and {Anderes}, E. and {Archipley}, M. and {Austermann}, J.~E. and {Avva}, J.~S. and {Aylor}, K. and {Balkenhol}, L. and {Barry}, P.~S. and {Thakur}, R. Basu and {Benabed}, K. and {Bianchini}, F. and {Bleem}, L.~E. and {Bouchet}, F.~R. and {Bryant}, L. and {Byrum}, K. and {Carlstrom}, J.~E. and {Carter}, F.~W. and {Cecil}, T.~W. and {Chang}, C.~L. and {Chaubal}, P. and {Chen}, G. and {Cho}, H. -M. and {Chou}, T. -L. and {Cliche}, J. -F. and {Crawford}, T.~M. and {Cukierman}, A. and {Daley}, C. and {de Haan}, T. and {Denison}, E.~V. and {Dibert}, K. and {Ding}, J. and {Dobbs}, M.~A. and {Everett}, W. and {Feng}, C. and {Ferguson}, K.~R. and {Fu}, J. and {Galli}, S. and {Gambrel}, A.~E. and {Gardner}, R.~W. and {Gualtieri}, R. and {Guns}, S. and {Gupta}, N. and {Guyser}, R. and {Halverson}, N.~W. and {Harke-Hosemann}, A.~H. and {Harrington}, N.~L. and {Henning}, J.~W. and {Hilton}, G.~C. and {Hivon}, E. and {Holder}, G.~P. and {Holzapfel}, W.~L. and {Hood}, J.~C. and {Howe}, D. and {Huang}, N. and {Irwin}, K.~D. and {Jeong}, O.~B. and {Jonas}, M. and {Jones}, A. and {Khaire}, T.~S. and {Knox}, L. and {Kofman}, A.~M. and {Korman}, M. and {Kubik}, D.~L. and {Kuhlmann}, S. and {Kuo}, C. -L. and {Lee}, A.~T. and {Leitch}, E.~M. and {Lowitz}, A.~E. and {Lu}, C. and {Meyer}, S.~S. and 
[…]
