### 2025-09-30 #8302 @astrorya
Hey @dstansby ! We would still like to proceed with the submission. We reduced the length by taking out the instructions, since we have a readthedocs available anyway, as well as fixed the issue with using NASA ADS for our bibliography per the comment you referenced. Let us know how we should proceed further.
