### 2025-10-26 #8302 @dstansby
👋 @ortk95, @cmillion, @ryanhausen - based on your previous experience in open source software development for astronomy, I'd like to invite you to review this Journal of Open Source Software paper - would you be interested? For more information on what reviewing for JOSS involves, see here: [joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) Any questions, let me know! The most recent version of the paper can be seen here: https://github.com/openjournals/joss-reviews/issues/8302#issuecomment-3393141060
