### 2025-12-23 #9686 @editorialbot
[REVIEW]: Image Marker
**Submitting author:** @bleeml (<a href="http://orcid.org/0000-0001-7665-5079">Lindsey Bleem</a>)
**Repository:** https://github.com/andikisare/imgmarker
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.9.2
