### 2026-01-26 #9686 @alexlyttle
@dstansby I am continuing with the review now, thanks for the reminder.

A quick question regarding this point:

> Contribution and authorship: Has the submitting author (bleeml) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

If measuring by commits, the submitting author has not made major contributions according to the [contributor stats](https://github.com/andikisare/imgmarker/graphs/contributors), but all authors have made contributions and the author list is appropriate. Since contributions to code development do not solely comprise commits, I don't see this as an issue. Please could you advise on this?
