### 2026-01-27 #9686 @dstansby
As you say, number of commits is just one factor to take into account on this point. I would recommend looking at the commits themselves and seeing if they are major changes to the software (e.g., bug fixes, doc improvements, code changes, new features vs. very minor changes). Does that help you to answer this point on the review checklist?
