### 2026-01-29 #9686 @bleeml
I can also add to this comment. I was the supervisor of the 2 students (Andi and Ryan) who did the majority of the coding (some pieces of code of mine from a private repo were folded into commits by them but they led by far the coding work). I submitted the paper to JOSS as Andi and Ryan were transitioning to graduate school and moving email contacts, etc. I am happy if either of them were considered the lead if that is a concern.
