### 2026-02-06 andikisare/imgmarker#32 @alexlyttle
JOSS review – overview of software documentation and functionality
The software functionality is clearly laid out in the paper and detailed documentation. Image Marker was easy to install and use. I can see this being useful in astronomy and other disciplines with a need to annotate image datasets.

I have raised a few minor issues relating to the documentation and functionality. I will follow with an issue containing my comments on the paper as soon as possible.

### Documentation

I suggest resolving these issues to help improve the documentation and mitigate confusion for new users.

+ [ ] #29
+ [ ] #30
+ [ ] #33

### Functionality

Resolving issue #31 would make the software easier to use but I do not think its resolution is required to accept the software and paper to JOSS. However, I suggest returning to and attempting to resolve existing issues labelled as bugs prior to acceptance.

+ [ ] #11
+ [ ] #20

### Questions

I have a few questions regarding the wider use of this software in the community.

+ Do you plan on increasing the available number of image categories and mark groups? It may be that the imposed limits on these reduce possible use cases. Customising such limits would be a nice addition in a future version of the software.
+ If not, can you give reasoning for the limits on image categories, mark groups, and maximum marks per group?
