### 2026-02-06 #9686 @alexlyttle
Thank you for the context @bleeml and apologies for the delay in providing a review.

I have created some issues in the source repository with my review of the software documentation and functionality. Issue https://github.com/andikisare/imgmarker/issues/32 provides an overview of these. My comments on the paper will follow.
