### 2026-02-18 #9686 @dstansby
👋 @bleeml I wanted to check in and see if you've had a chance to look at the review comments in https://github.com/andikisare/imgmarker/issues/32?

@enourbakhsh it would be great if you could finish your review soon. Let me know if you have any questions or issues.

@alexlyttle just to check, is your first review complete and you can re-review once your comments have been addressed?
