### 2026-03-21 andikisare/imgmarker#34 @enourbakhsh
JOSS review - crash on cancelled/empty directory selection
When launching `imgmarker` executable, if you cancel the directory picker or pick a folder with no image files in it, the app crashes instead of showing an error or re-prompting:

```python
File ".../imgmarker/gui/window.py", line 919, in __init_data__
    self.images, self.idx = io.glob(edited_images=self.images)
File ".../imgmarker/io.py", line 495, in glob
    paths = sorted(_glob.glob(os.path.join(config.IMAGE_DIR, '*.*')))
TypeError: expected str, bytes or os.PathLike object, not NoneType

    self.image = self.images[self.idx]
IndexError: list index out of range
```

Looks like `config.IMAGE_DIR` ends up as `None` and nothing catches that. Would be nice to get a dialog like "No images found, please choose another directory" instead of a traceback.

Tested using:
- imgmarker 1.1.4
- Python 3.12.11 (conda)
- macOS x86_64
