### 2026-03-21 andikisare/imgmarker#35 @enourbakhsh
JOSS review - mark group selection isn't obvious in the UI
First time using the tool, I spent a while clicking around, and all my marks ended up as group `1`. Without reading the docs, it wasn't clear/intuitive that I needed to press a number key to switch groups. I realized left-click just silently defaults to group 1 without any indication in the UI itself.

A few related things:

1. **No active group indicator:**
After pressing e.g. `2` to switch to group 2, there's nothing on screen confirming which group I'm now placing marks in. A label in the status bar or near the cursor would help.

2. **Category checkboxes are unlabeled:**
The 1-5 checkboxes and heart icon at the bottom of the window are image-level categories, but there's no tooltip or label explaining what they do or how they're different from mark groups.

3. **Things like `MiddleMouseButton` etc. not available on laptop touchpads / Magic Mouse:**
As an example, the controls reference `MiddleMouseButton` for focus, but laptop touchpads (Windows/Linux) and Mac trackpads or Magic Mice don’t have a dedicated middle button. It might be worth providing an alternative shortcut or noting this limitation in the docs.

Just some minor issues I noticed, but confusing on first use.

Tested using:
- imgmarker 1.1.4
- macOS x86_64
