### 2026-03-21 #9686 @enourbakhsh
Sorry for the delayed review. I installed Image Marker v1.1.4 on my Intel macOS (Python 3.12, conda) and tested it with the bundled SPT cluster TIFF cutouts. The first thing I noticed was the nice WCS readout on hover. The multi-frame FITS/TIFF support is a great addition, and the Gaussian blur helped with pulling out faint features in the SPT cutouts I used in my tests. I spent some more time with the tool and am sharing my findings here.

I've opened a couple of issues in the repo, but you might need to open more based on my review:

- andikisare/imgmarker#34
- andikisare/imgmarker#35

### Installation

My first attempt was `pip install imgmarker` in a Python 3.7 conda env (my default). `pip` spent a few minutes backtracking through every version on PyPI and eventually failed with a confusing astropy dependency error. The actual problem is that PyQt6 and astropy need `Python >= 3.9`, but your `pyproject.toml` doesn't set `requires-python`, so pip has no way to know upfront and just fails. Maybe adding `requires-python = ">=3.9"` is a good way to save users from that. Once I made a fresh 3.12 env it installed fine.

### Robustness and usability

First time launching, I think I either cancelled the directory picker or chose a folder with no images in it. Either way, the app just crashed with `TypeError` on `config.IMAGE_DIR` being `None`, then an `IndexError`. I didn't see any error dialog or re-prompt, just a traceback. This would be the very first thing a new user runs into, so it should be handled more gracefully. Details in the issue andikisare/imgmarker#34.

Also, took me a bit to figure out mark groups without relying too much on the docs. I kept clicking, and everything showed up as group `1`. I now know left-click defaults to group 1, and you have to press a number key to switch groups, but there's nothing in the UI that tells you which group is active, so it's easy to not even realize groups are a thing. The category checkboxes at the bottom (1-5 plus a heart/favorite) were also a bit of a mystery to me until I dug into the docs. Some tooltips or a status bar indicator would help a lot here. See the issue andikisare/imgmarker#35.

The comment text box was also slightly confusing. There's an `Enter` button next to it, but it turns out comments auto-save when you navigate to the next image anyway. So it's not clear what `Enter` actually does that navigating away doesn't. Might be worth clarifying this in the UI or docs.

On the topic of Mac and touchpad support, the Controls window (F1) and the online docs both list things like _MiddleMouseButton_. That doesn't exist on touchpads, trackpads, or Magic Mice. I'd suggest either providing a Mac-friendly alternative (or at least noting that some controls assume a three-button mouse). I believe a good chunk of your target audience is on Macs.

One more thing, I remember seeing Ctrl instead of Cmd on my Intel Mac. I know there's [this commit](http://github.com/andikisare/imgmarker/commit/55ada779ff67aa653fc7a5ded15d64ddb21c224d) just for that, but please double check.

### Software paper

To start, the "statement of need" does a decent job positioning `Image Marker` relative to Zooniverse, FitsMap, DS9, and DES Exposure Checker. Some thoughts after reading the paper:

1. The paper talks about crowdsourcing within smaller collaborations (multiple people working on the same dataset), but each user just gets their own CSV, right? There's no explanation of how those individual results get merged or how agreement between users is measured, nothing like that. For the BCG centering use case where multiple people are classifying the same clusters, it would be good to at least mention how results get combined downstream, even if that happens outside the tool. Or maybe just acknowledge that, and the user would understand that further explanation is beyond the scope of your short paper.

2. I think the comparison to FitsMap is a little off. The paper says FitsMap lacks "methods for crowdsourcing" but Image Marker doesn't really have those either. I know it does support _basic_ multi-user file sharing, but it lacks things like aggregation and consensus, e.g., it's not clear how it handles conflicting marks from different users. Simply put, Image Marker is for flipping through lots of small cutouts quickly, FitsMap is for panning around large mosaics with catalog overlays. That distinction is clearer and more useful to the reader, IMO. You could point out basic multi-user file sharing in Image Marker, but clarify that it's not full crowdsourcing in a real sense.

3. The 8-bit per channel RGB limit is mentioned, but it's easy to get confused here. I know the paper notes: "Images that exceed these limitations will have their bit depth lowered" but it's not clearly stated what "lowered" means. What exactly happens when you load a 16-bit image? truncated, rescaled (like stretching 0-65535 to 0-255), etc.? Would there be any warnings? This should probably be documented more clearly since people will run into it.

4. Output is CSV only. Would be nice to have ECSV or VOTable as options. Most astro pipelines are going to want to read these into astropy Tables, and having proper column metadata would help with that. Ignore this if it's too much work, though.

5. On WCS: does it handle SIP distortions, TPV, etc., or just standard TAN? I'm guessing `astropy.wcs` takes care of it, but for survey data like DECam or SPT-3G, where more complex distortion terms are common, it'd be reassuring to mention what we're dealing with explicitly in the paper.

### CI

CI currently runs on Ubuntu only with Python 3.12, right? Since you provide macOS and Windows executables, adding at least one runner for each platform would be reassuring. A second Python version (3.10 or 3.11) wouldn't hurt either, especially with the missing `requires-python`.

### License

GitHub can't identify the license (the badge says "not identifiable") even though it's MIT. You might want to check this yourself, but it looks like the non-standard formatting (UChicago Argonne header, a `·` character instead of a dash) is throwing off GitHub's detection? Also, the file starts with "All Rights Reserved" right above the MIT text, which is a bit contradictory. See if using the standard MIT template or adding `license = "MIT"` in `pyproject.toml` as an SPDX identifier fixes this.

### Other minor things I noticed

- The paper says you provide Mac executables, but the releases only have ARM versions (`mac-arm.zip`). No dedicated Intel Mac executable? Probably not a big deal, but that's a gap for anyone still on an older Intel Mac. Maybe clarify it in the docs if you're not intending to fully support Intel Macs.
- It'd be nice to have `CONTRIBUTING.md` or `CITATION.cff` in the repo. I don't think they're blocking issues for JOSS acceptance, though.
- `pyproject.toml` lists two authors (Kisare, Walker) but the paper has three (Walker, Kisare, Bleem). Bleem is missing from the package metadata
