### 2026-05-29 bc07a89
Re-added automatically opening the keybinds window when the images first appear
