### 2026-05-29 55f55ee
Fixed application window showing as __main__.py instead of Image Marker, also attached version number, updated version to 1.1.5
