+++ imgmarker/__main__.py
@@ -6,12 +6,14 @@ Full license found at _YOUR_INSTALLATION_DIRECTORY_/imgmarker/LICENSE
 
 from imgmarker.gui.pyqt import QApplication, QIcon
 from imgmarker.gui.window import MainWindow, _open_save
-from imgmarker import config, ICON
+from imgmarker import config, ICON, __version__
 import sys
+
 
 def run():
     app = QApplication(sys.argv)
     app.setWindowIcon(QIcon(ICON))
+    app.setApplicationName(f"Image Marker v. {__version__}")
     
     config.SAVE_DIR = _open_save()
     config.IMAGE_DIR, config.GROUP_NAMES, config.CATEGORY_NAMES, config.GROUP_MAX, config.RANDOMIZE_ORDER = config.read()
