### 2026-06-13 26a0a80
Added notes to settings window, settings pops up by default, changed default num max marks to -1, -1 for no limit, updated version to 1.1.6
