+++ imgmarker/gui/window.py
@@ -94,9 +94,8 @@ class SettingsWindow(QWidget):
             spinbox.setFixedHeight(30)
             spinbox.setMaximum(9)
             spinbox.setMinimum(-1)
-            spinbox.setValue("-1")
-            value:str = config.GROUP_MAX[i]
-            print(value, type(value))
+            spinbox.setValue(-1)
+            value:str = str(config.GROUP_MAX[i])
             if value.isnumeric(): spinbox.setValue(int(value))
             spinbox.valueChanged.connect(self.update_config)
             self.max_boxes.append(spinbox)
