### 2026-08-19 #9686 @alexlyttle
Thanks @bleeml, @enourbakhsh, and @astrorya for addressing issues raised in the code repository. I have installed the application to verify the changes and reviewed the resolutions to issues raised.

I am happy with the changes made to the paper and software. I have no additional comments about the paper to add to @enourbakhsh's review and your response. The paper makes the scope and usage of the software clear. Also, I can see that you have addressed several other issues raised during review and resolved existing bugs. The software is in an excellent state for publication and I look forward to seeing how it evolves and gets used in the future.

@dstansby I will also gladly recommend acceptance to JOSS.
