### 2026-08-30 #9686 @bleeml
@dstansby

Double check authors and affiliations (including ORCIDs)
	- These are all correct.

Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
      - Version 1.2.0 (current version) is correct

Archive the release on Zenodo/figshare/etc and post the DOI here.
We were given a DOI for this on DOECODE
https://www.osti.gov/doecode/biblio/158455

Is this sufficient?   Or should we also add to Zenodo?

Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
Done

Make sure that the license listed for the archive is the same as the software license.
Done (on doecode matches MIT license, on the arxiv the paper is CC-4, will update paper to JOSS version when accepted)
