### 2026-09-01 #9686 @dstansby
> Is this sufficient? Or should we also add to Zenodo?

Is that an archive of the software, or just a DOI? I can't seem to find a way to download a specific version of the code directly from https://www.osti.gov/doecode/biblio/158455. If it's just a DOI, it will need archiving somewhere else that assigns a DOI and archives the software - most folks use Zenodo.

I'm just going through the paper now, will get back to you shortly with my editor corrections.
