### 2026-09-01 #9686 @dstansby
@bleeml I've opened a PR with minor copy edits to the paper.

In addition, the paper is currently slightly on the long side for JOSS. Because papers in JOSS are targeted towards a wide audience, please could you cut down the "Our use case" paragraph. I'm happy having a brief discussion of your use case, but it's currently quite technical, and could be slimmed down and with fewer references.

Finally, on the topic of references please
- Fix the PyQT documentation author, currently it lists "Limited, R. C.". I suspect it should have no authors?
- Add the DOI to "An image is worth 16x16 words..."
