### 2026-09-04 #9686 @astrorya
Hi @dstansby , we have:

- trimmed the "Our Use Case" section by ~85 words
- fixed the author name for PyQT
- added the DOI for "An image is worth 16x16 words..." (though it got cut from that paragraph in the end)
- made a Zenodo page here: https://doi.org/10.5281/zenodo.22260997

And as you already saw, we also accepted your push request with a small addition to dissolve ambiguity.
